+++ src/config/polaris-server-config.ts
@@ -4,12 +4,12 @@ import {
     ApplicationProperties,
     LoggerConfiguration,
 } from '@enigmatis/polaris-logs';
-import { PolarisConnection } from '@enigmatis/polaris-typeorm';
 import { ApolloServerExpressConfig } from 'apollo-server-express';
 import { DocumentNode } from 'graphql';
 import { IResolvers } from 'graphql-tools';
 import { ExpressContext } from '..';
 import { MiddlewareConfiguration } from '../index';
+import { SnapshotConfiguration } from './snapshot-configuration';
 
 export interface PolarisServerConfig extends ApolloServerExpressConfig {
     typeDefs: DocumentNode | DocumentNode[] | string | string[];
@@ -21,8 +21,8 @@ export interface PolarisServerConfig extends ApolloServerExpressConfig {
     allowSubscription: boolean;
     customMiddlewares?: any[];
     customContext?: (context: ExpressContext) => any;
-    connection?: PolarisConnection;
     supportedRealities: RealitiesHolder;
     shouldAddWarningsToExtensions: boolean;
     allowMandatoryHeaders: boolean;
+    snapshotConfig: SnapshotConfiguration;
 }