+++ src/config/snapshot-configuration.ts
@@ -1,0 +1,7 @@
+export interface SnapshotConfiguration {
+    snapshotCleaningInterval: number;
+    secondsToBeOutdated: number;
+    maxPageSize: number;
+    entitiesAmountPerFetch: number;
+    autoSnapshot: boolean;
+}