+++ src/index.ts
@@ -4,6 +4,7 @@ export * from './config/create-apollo-config-util';
 export { PolarisServerOptions } from './config/polaris-server-options';
 export { MiddlewareConfiguration } from './config/middleware-configuration';
 export { polarisFormatError } from './errors/error-formatter';
+export { SnapshotConfiguration } from './config/snapshot-configuration';
 export * from 'apollo-server-errors';
 export { PubSub, gql } from 'apollo-server-express';
 export * from '@enigmatis/polaris-common';
@@ -15,3 +16,12 @@ export * from '@enigmatis/polaris-schema';
 export * from '@enigmatis/polaris-typeorm';
 export { Entity } from '@enigmatis/polaris-typeorm';
 export { getPolarisServerConfigFromOptions } from './server/configurations-manager';
+export {
+    IResolvers,
+    IResolverObject,
+    IResolverOptions,
+    IResolversParameter,
+    IResolverValidationOptions,
+    IFieldResolver,
+} from 'graphql-tools';
+export * from './snapshot/paginated-resolver';