+++ src/server/configurations-manager.ts
@@ -1,6 +1,6 @@
 import { RealitiesHolder } from '@enigmatis/polaris-common';
 import { LoggerConfiguration } from '@enigmatis/polaris-logs';
-import { MiddlewareConfiguration, PolarisServerOptions } from '..';
+import { MiddlewareConfiguration, PolarisServerOptions, SnapshotConfiguration } from '..';
 import { PolarisServerConfig } from '../config/polaris-server-config';
 
 export const getDefaultMiddlewareConfiguration = (): MiddlewareConfiguration => {
@@ -8,6 +8,7 @@ export const getDefaultMiddlewareConfiguration = (): MiddlewareConfiguration => 
         allowDataVersionAndIrrelevantEntitiesMiddleware: true,
         allowRealityMiddleware: true,
         allowSoftDeleteMiddleware: true,
+        allowTransactionalMutations: true,
     };
 };
 
@@ -16,6 +17,16 @@ export const getDefaultLoggerConfiguration = (): LoggerConfiguration => {
         loggerLevel: 'info',
         writeToConsole: true,
         writeFullMessageToConsole: false,
+    };
+};
+
+export const getDefaultSnapshotConfiguration = (): SnapshotConfiguration => {
+    return {
+        snapshotCleaningInterval: 60,
+        secondsToBeOutdated: 60,
+        maxPageSize: 50,
+        entitiesAmountPerFetch: 50,
+        autoSnapshot: false,
     };
 };
 
@@ -51,5 +62,6 @@ export const getPolarisServerConfigFromOptions = (
                 : options.shouldAddWarningsToExtensions,
         allowMandatoryHeaders: options.allowMandatoryHeaders || false,
         supportedRealities: getSupportedRealities(options),
+        snapshotConfig: options.snapshotConfig || getDefaultSnapshotConfiguration(),
     };
 };